app/requirements.txt
adaptivecardbuilder==0.0.9
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
coloredlogs==15.0.1
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
frozenlist==1.4.0
future==0.18.3
greenlet==2.0.2
humanfriendly==10.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.9
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
requests-toolbelt==1.0.0
SQLAlchemy==2.0.20
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
webex-bot==0.4.0
webexteamssdk==1.6.1
websockets==10.2
Werkzeug==2.3.7
yarl==1.9.2
